progress on actions interface
